requirements.txt
numpy
scipy
tqdm
pandas
matplotlib
sklearn
jupyterlab
ipywidgets
scikit-learn
scikit-image
scikit-optimize
seaborn
pydicom
latex
gsppy
torchsummary
tensorboard
tensorboardX
Pillow
grad-cam
blobfile
visdom
nibabel
imageio